fix(ios): avoid boolean/string crash by disabling screens
