+++ App.jsx
@@ -1,0 +1,36 @@
+import "react-native-gesture-handler";
+import { enableScreens } from "react-native-screens";
+import { useEffect } from "react";
+import { Platform } from "react-native";
+import { StatusBar } from "expo-status-bar";
+import { NavigationContainer } from "@react-navigation/native";
+import { GestureHandlerRootView } from "react-native-gesture-handler";
+import { SafeAreaProvider } from "react-native-safe-area-context";
+import { AuthNavigator } from "./src/navigation/AuthNavigator";
+import { ThemeProvider } from "./src/context/ThemeContext";
+import { AuthProvider } from "./src/context/AuthContext";
+
+enableScreens(false);
+
+export default function App() {
+  useEffect(() => {
+    if (Platform.OS === "web") {
+      document.body.style.overflow = "auto";
+    }
+  }, []);
+  return (
+    <GestureHandlerRootView style={{ flex: 1 }}>
+      <SafeAreaProvider>
+        <ThemeProvider>
+          <AuthProvider>
+            <NavigationContainer>
+              <AuthNavigator />
+              <StatusBar style="light" />
+            </NavigationContainer>
+          </AuthProvider>
+        </ThemeProvider>
+      </SafeAreaProvider>
+    </GestureHandlerRootView>
+  );
+}
+
